Ajusta pagamento ao usuario

+++ database/migrations/2020_10_20_005126_create_transactions_table.php
@@ -15,7 +15,7 @@ class CreateTransactionsTable extends Migration
     {
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
-            $table->float('transaction_value');
+            $table->double('transaction_value');
             $table->unsignedBigInteger('credit_card_id');
             $table->foreign('credit_card_id')
                 ->references('id')
